+++ resolwe/elastic/management/commands/elastic_index.py
@@ -24,8 +24,10 @@ class Command(ElasticIndexFilterMixin, BaseCommand):
 
     def handle(self, *args, **options):
         """Command handle."""
+        verbosity = int(options['verbosity'])
+
         if self.has_filter(options):
-            self.filter_indices(options)
+            self.filter_indices(options, verbosity)
         else:
             # Process all indices.
             index_builder.build()